feat: adapting the runtime to support adapter work (PL-32)
