+++ lib/services/runtime/handlers/caroussel/carousel.google.ts
@@ -7,7 +7,7 @@ import { VoiceflowConstants } from '@voiceflow/voiceflow-types';
 
 import { HandlerFactory } from '@/runtime';
 
-import { isGooglePlatform } from '../../utils.google';
+import { isGooglePlatform } from '../../utils';
 import { NoReplyGoogleHandler } from '../noReply/noReply.google';
 import { CarouselHandler, handlerUtils } from './carousel';
