+++ package.json
@@ -145,7 +145,7 @@
     "build": "ttsc --project ./tsconfig.build.json && cpy app.config.js yarn.lock build/",
     "commit": "cz",
     "connect:cli": "vfcli mesh connect general-runtime -c \"/bin/bash\"",
-    "dev": "ts-node-dev --files --respawn --exit-child --transpile-only -r tsconfig-paths/register -- start.ts",
+    "dev": "ts-node-dev --files --respawn --exit-child --transpile-only --inspect -r tsconfig-paths/register -- start.ts",
     "docs:gen": "typedoc ./lib ./config.ts",
     "e2e": "yarn kill:e2e && NODE_TLS_REJECT_UNAUTHORIZED=0 NODE_ENV=e2e yarn start:base",
     "e2e:check-deps": "npx wait-on https://localhost:8011/health",
